Created game component to use for creating and edit games
